Batch consistency via configuration.  Fix test.
