+++ src/test/java/com/hmsonline/trident/cql/ConditionalUpdateTest.java
@@ -8,10 +8,6 @@ import static com.hmsonline.trident.cql.incremental.example.SalesAnalyticsMapper
 import static com.hmsonline.trident.cql.incremental.example.SalesAnalyticsMapper.KEY_NAME;
 import static com.hmsonline.trident.cql.incremental.example.SalesAnalyticsMapper.TABLE_NAME;
 import static com.hmsonline.trident.cql.incremental.example.SalesAnalyticsMapper.VALUE_NAME;
-import static org.junit.Assert.assertEquals;
-
-import java.util.HashMap;
-import java.util.Map;
 
 import org.junit.Ignore;
 import org.junit.Test;
@@ -20,10 +16,6 @@ import org.junit.runners.JUnit4;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import com.datastax.driver.core.ResultSet;
-import com.datastax.driver.core.Row;
-import com.datastax.driver.core.Statement;
-import com.datastax.driver.core.querybuilder.Select;
 import com.datastax.driver.core.querybuilder.Update;
 
 /**
@@ -31,32 +23,12 @@ import com.datastax.driver.core.querybuilder.Update;
  */
 @Ignore
 @RunWith(JUnit4.class)
-public class ConditionalUpdateTest {
+public class ConditionalUpdateTest extends StateTest {
     private static final Logger LOG = LoggerFactory.getLogger(ConditionalUpdateTest.class);
-    public CqlClientFactory clientFactory;
-    public Map<String, String> configuration = new HashMap<String, String>();
     public String APPLIED_COLUMN = "[applied]";
 
     public ConditionalUpdateTest() {
-        configuration.put(MapConfiguredCqlClientFactory.TRIDENT_CASSANDRA_CQL_HOSTS, "localhost");
-        clientFactory = new MapConfiguredCqlClientFactory(configuration);
-    }
-
-    public void assertValue(String k, Integer expectedValue) {
-        Select selectStatement = select().column("v").from(KEYSPACE_NAME, TABLE_NAME);
-        selectStatement.where(eq(KEY_NAME, k));
-        ResultSet results = clientFactory.getSession().execute(selectStatement);
-        Integer actualValue = results.one().getInt(VALUE_NAME);
-        assertEquals(expectedValue, actualValue);
-    }
-
-    public void executeAndAssert(Statement statement, String k, Integer expectedValue) {
-        LOG.debug("EXECUTING [{}]", statement.toString());
-        ResultSet results = clientFactory.getSession().execute(statement);
-        Row row = results.one();
-        if (row != null)
-            LOG.debug("APPLIED?[{}]", row.getBool("[applied]"));
-        this.assertValue(k, expectedValue);
+        super();
     }
 
     @Test
